refactor(columnar): move slow expression evaluation
